cleanup unit tests
Reorganize tests so each test file only tests functions from its
corresponding source file:

- Remove redundant formatDefaultOutput tests from top_test.go (already
  covered in output_test.go)
- Remove duplicate isWatchMode test from events_test.go (already in
  get_test.go)
- Move formatLogsOutput tests from logs_test.go to output_test.go
- Move formatOutput api-versions/api-resources tests from
  api_versions_test.go to output_test.go
- Move streamLines/streamLinesFilterHeader tests from logs_test.go and
  get_test.go to new exec_test.go
- Remove unused captureOutputAndStderr helper
- Update AGENTS.md with unit test scoping guidelines
